Refactor and fix a duplication enum for DepartmentStatusEnum

+++ src/main/java/com/youhujia/solar/domain/department/DepartmentDAO.java
@@ -13,25 +13,26 @@ import java.util.List;
  */
 @Transactional
 public interface DepartmentDAO extends JpaRepository<Department, Long> {
+
     List<Department> findByOrganizationIdAndGuest(Long organizationId, Boolean isGuest);
 
-    List<Department> findByOrganizationIdAndGuestAndStatus(Long organizationId, Boolean isGuest, Byte status);
+    List<Department> findByOrganizationIdAndGuestAndStatus(Long organizationId, Boolean isGuest, Integer status);
 
     List<Department> findByOrganizationId(Long organizationId);
 
     List<Department> findByHostId(Long departmentId);
 
-    List<Department> findByHostIdAndStatus(Long departmentId, Byte status);
+    List<Department> findByHostIdAndStatus(Long departmentId, Integer status);
 
     List<Department> findByNumber(String dptNum);
 
-    List<Department> findByNumberAndStatus(String dptNum, Byte status);
+    List<Department> findByNumberAndStatus(String dptNum, Integer status);
 
-    List<Department> findByOrganizationIdAndStatus(Long organizationId, Byte status);
+    List<Department> findByOrganizationIdAndStatus(Long organizationId, Integer status);
 
     @Modifying
     @Query(value = "from Department d where d.organizationId = ?1 and d.status > ?2")
-    List<Department> findByOrganizationIdWithStatus(Long orgId, Byte status);
+    List<Department> findByOrganizationIdWithStatus(Long orgId, Integer status);
 
     @Modifying
     @Query(value = "select d from Department d where d.guest = false and d.organizationId in :ids")
